Implementation change for support of SYMO GEN24
As currently the SYMO GEN24 is not supported, a large rework was needed. There are many parameters changed on the solarAPI. Therefore changed the creation and filling of objects to a dynamic way instead of static. Still static definitions for objects used as the settings of the objects cannot automized. All not defined objects will created with default settings.
Implementation is not finished as the storage is still missing